update clear all

+++ clear_all.py
@@ -10,12 +10,14 @@ os.makedirs('./log/filtrate_log')
 os.makedirs('./log/render_log')
 os.makedirs('./log/glb2mesh_log')
 
-folder_path = './glb_data'
-for filename in os.listdir(folder_path):
-    if 'gitkeep' in filename:
-        continue
-    file_path = os.path.join(folder_path, filename)
-    if os.path.isdir(file_path):
-        shutil.rmtree(file_path)  # 删除子文件夹及其内容
-    else:
-        os.remove(file_path)  # 删除文件
+clear_glb_data = True
+if clear_glb_data:
+    folder_path = './glb_data'
+    for filename in os.listdir(folder_path):
+        if 'gitkeep' in filename:
+            continue
+        file_path = os.path.join(folder_path, filename)
+        if os.path.isdir(file_path):
+            shutil.rmtree(file_path)  # 删除子文件夹及其内容
+        else:
+            os.remove(file_path)  # 删除文件
